+++ app/Jobs/ExecuteCoolifyProcess.php
@@ -2,9 +2,8 @@
 
 namespace App\Jobs;
 
-use App\Services\RemoteProcess\RemoteProcess;
+use App\Actions\RemoteProcess\RemoteProcess;
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
